update file task 10-div.sql

+++ MySQL_Advanced/10-div.sql
@@ -1,9 +1,9 @@
--- 10. Safe divide
+-- path: 10-div.sql
 -- creates a function SafeDiv that divides (and returns) the first by the second number or returns 0 if the second number is equal to 0
 DROP FUNCTION IF EXISTS SafeDiv;
 DELIMITER $$
-CREATE FUNCTION SafeDiv(a INT, b INT) RETURNS DECIMAL (10, 2)
-DETERMINISTIC
+CREATE FUNCTION SafeDiv(a INT, b INT)
+RETURNS FLOAT DETERMINISTIC
 BEGIN
     IF (b = 0)
     THEN
